FormFillingActivity: show cause type

+++ app/src/main/java/com/czttgd/android/zhijian/data/Inspection.kt
@@ -106,7 +106,7 @@ object Inspection {
         }.parseResponse<Long>().data!!
     }
 
-    suspend fun update(record: InspectionForm, id: Int) {
+    suspend fun update(record: InspectionForm, id: Long) {
         appHttpClient.put("$serverAddr/inspection/$id") {
             contentType(ContentType.Application.FormUrlEncoded)
             setFormDataBody(record)
